pyproject.toml
[tool.poetry]
name = "nonebot_plugin_mute10rolls"
version = "0.1.0a2"
description = "10 rolls for self-muted!"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.2"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-apscheduler = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencentyun"
url = "https://mirrors.tencent.com/pypi/simple"
default = true